refactor(db): move repository tests from mod.rs to db/repository.rs
