+++ src/db/mod.rs
@@ -298,63 +298,7 @@ fn is_already_exists_error(err: &rusqlite::Error) -> bool {
 mod tests {
     use super::*;
     use crate::db::test_support::*;
-    use crate::model::{ChatRole, Repository};
-
-    // --- Repository settings tests ---
-
-    #[test]
-    fn test_update_repository_name() {
-        let db = Database::open_in_memory().unwrap();
-        db.insert_repository(&make_repo("r1", "/tmp/repo1", "repo1"))
-            .unwrap();
-        db.update_repository_name("r1", "My Custom Name").unwrap();
-        let repos = db.list_repositories().unwrap();
-        assert_eq!(repos[0].name, "My Custom Name");
-        // path_slug should remain unchanged
-        assert_eq!(repos[0].path_slug, "repo1");
-    }
-
-    #[test]
-    fn test_update_repository_icon() {
-        let db = Database::open_in_memory().unwrap();
-        db.insert_repository(&make_repo("r1", "/tmp/repo1", "repo1"))
-            .unwrap();
-
-        // Set icon
-        db.update_repository_icon("r1", Some("rocket")).unwrap();
-        let repos = db.list_repositories().unwrap();
-        assert_eq!(repos[0].icon.as_deref(), Some("rocket"));
-
-        // Clear icon
-        db.update_repository_icon("r1", None).unwrap();
-        let repos = db.list_repositories().unwrap();
-        assert!(repos[0].icon.is_none());
-    }
-
-    #[test]
-    fn test_repository_path_slug_persisted() {
-        let db = Database::open_in_memory().unwrap();
-        let repo = Repository {
-            id: "r1".into(),
-            path: "/tmp/my-project".into(),
-            name: "My Project".into(),
-            path_slug: "my-project".into(),
-            icon: None,
-            created_at: String::new(),
-            setup_script: None,
-            custom_instructions: None,
-            sort_order: 0,
-            branch_rename_preferences: None,
-            setup_script_auto_run: false,
-            base_branch: None,
-            default_remote: None,
-            path_valid: true,
-        };
-        db.insert_repository(&repo).unwrap();
-        let repos = db.list_repositories().unwrap();
-        assert_eq!(repos[0].name, "My Project");
-        assert_eq!(repos[0].path_slug, "my-project");
-    }
+    use crate::model::ChatRole;
 
     #[test]
     fn test_last_message_per_workspace() {
